ENH: Support validation data set in model training
Add the ability to specify a subset of the data for validation. These
validation feature vectors are a holdout dataset: they will have their
loss and accuracy statistics separately tracked and they will be
ineligible for selection in the active learning process.

Add some API functions:

    get_training_feature_vectors
    get_validation_feature_vectors
    get_training_labels
    get_validation_labels
    set_validation_indices
    get_validation_indices
    clear_validation_indices

Change the names of some API functions to clarify that they are
handling, e.g., a list of feature vectors rather than a list of
individual feature values:

    read_all_features_from_h5py --> read_all_feature_vectors_from_h5py
    write_all_features_from_h5py --> write_all_feature_vectors_from_h5py
    set_all_features --> set_all_feature_vectors
    get_all_features --> get_all_feature_vectors
    clear_all_features --> clear_all_feature_vectors
    set_some_features --> set_some_feature_vectors
    get_some_features --> get_some_feature_vectors